requirements.txt
kagglehub
pandas
seaborn
ipykernel
statsmodels
PyYAML
sklearn